Del my old files, replaced refactored version of data_prep in original, incporated Tiffany comment, redid analysis
