--- 5311/Crosswalks/vehicles-left-only.csv
@@ -1,10 +1,0 @@
-﻿agency,ntd_id,itp_id
-San Benito County LTA,9R02-91009,274
-County of Sacramento Municipal Services Agency Department of Transportation,9R02-90216,
-"City of Dixon, dba: Readi-Ride",9R02-91041,94
-"Mariposa County Transit, dba: Mari-Go",9R02-91082,
-Calaveras Transit Agency,9R02-99442,
-Palo Verde Valley Transit Agency,9R02-99454,238
-"City of Ridgecrest, dba: Ridgerunner",9R02-91006,263
-Modoc Transportation Agency,9R02-91008,
-"Wasco, City of",9R02-99426,363
